requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-cors-headers
sentry-sdk[django]
django-otp 
django-two-factor-auth 
qrcode